SelectedAllCheckbox -> SelectedCheckboxCell
