--- src/NewDataTable/Components/SelectAllCheckbox/index.ts
@@ -1,1 +1,0 @@
-export { default as SelectedAllCheckbox } from "./SelectedAllCheckbox"
