+++ src/NewDataTable/Components/SelectedCheckboxCell/index.ts
@@ -1,0 +1,1 @@
+export { default as SelectedCheckboxCell } from "./SelectedCheckboxCell"
